--- test/odbc/psqlodbc/test_pg_datatype_money.cpp
@@ -1,913 +1,0 @@
-#include <gtest/gtest.h>
-#include <sqlext.h>
-#include "../src/odbc_handler.h"
-#include "../src/query_generator.h"
-#include "../src/drivers.h"
-#include <cmath>
-#include <iostream>
-using std::pair;
-
-const string TABLE_NAME = "master_dbo.money_table_odbc_test";
-const string COL1_NAME = "pk";
-const string COL2_NAME = "data";
-const string DATATYPE_NAME = "sys.money";
-const string VIEW_NAME = "master_dbo.money_view_odbc_test";
-vector<pair<string, string>> TABLE_COLUMNS = {
-  {COL1_NAME, "INT PRIMARY KEY"},
-  {COL2_NAME, DATATYPE_NAME}
-};
-const int DATA_COLUMN = 2;
-const int BUFFER_SIZE = 256;
-const int INT_BYTES_EXPECTED = 4;
-const int DOUBLE_BYTES_EXPECTED = 8;
-
-class PSQL_DataTypes_Money : public testing::Test {
-  void SetUp() override {
-    if (!Drivers::DriverExists(ServerType::PSQL)) {
-      GTEST_SKIP() << "PSQL Driver not present: skipping all tests for this fixture.";
-    }
-
-    OdbcHandler test_setup(Drivers::GetDriver(ServerType::PSQL));
-    test_setup.ConnectAndExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-  }
-
-  void TearDown() override {
-    OdbcHandler test_teardown(Drivers::GetDriver(ServerType::PSQL));
-    test_teardown.ConnectAndExecQuery(DropObjectStatement("VIEW", VIEW_NAME));
-    test_teardown.CloseStmt();
-    test_teardown.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-  }
-};
-
-// Helper to convert string to double
-double StringToDouble(const string &value) {
-  return std::stod(value);
-}
-
-TEST_F(PSQL_DataTypes_Money, Table_Creation) {
-  // TODO - Expected needs to be fixed.
-  const int LENGTH_EXPECTED = 255;
-  const int PRECISION_EXPECTED = 0;
-  const int SCALE_EXPECTED = 0;
-  const string NAME_EXPECTED = "unknown";
-
-  char name[BUFFER_SIZE];
-  SQLLEN length;
-  SQLLEN precision;
-  SQLLEN scale;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  // Create a table with columns defined with the specific datatype being tested.
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Select * From Table to ensure that it exists
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-
-  // Make sure column attributes are correct
-  rcode = SQLColAttribute(odbcHandler.GetStatementHandle(),
-                          DATA_COLUMN,
-                          SQL_DESC_LENGTH, // Get the length of the column (size of char in columns)
-                          NULL,
-                          0,
-                          NULL,
-                          (SQLLEN *)&length);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(length, LENGTH_EXPECTED);
-
-  rcode = SQLColAttribute(odbcHandler.GetStatementHandle(),
-                          DATA_COLUMN,
-                          SQL_DESC_PRECISION, // Get the precision of the column
-                          NULL,
-                          0,
-                          NULL,
-                          (SQLLEN *)&precision);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(precision, PRECISION_EXPECTED);
-
-  rcode = SQLColAttribute(odbcHandler.GetStatementHandle(),
-                          DATA_COLUMN,
-                          SQL_DESC_SCALE, // Get the scale of the column
-                          NULL,
-                          0,
-                          NULL,
-                          (SQLLEN *)&scale);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(scale, SCALE_EXPECTED);
-
-  rcode = SQLColAttribute(odbcHandler.GetStatementHandle(),
-                          DATA_COLUMN,
-                          SQL_DESC_TYPE_NAME, // Get the type name of the column
-                          name,
-                          BUFFER_SIZE,
-                          NULL,
-                          NULL);
-  ASSERT_EQ(string(name), NAME_EXPECTED);
-
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_Money, Insertion_Success) {
-  int pk;
-  double data;
-  SQLLEN pk_len;
-  SQLLEN data_len;
-  SQLLEN affected_rows;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<string> INSERTED_DATA = {
-    "-922337203685477.5808",
-    "922337203685477.5807",
-    "0",
-    "100000000.01",
-    "-100000000.01",
-    "NULL"
-  };
-  const int NUM_OF_INSERTS = INSERTED_DATA.size();
-
-  const vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {
-    {1, SQL_C_LONG, &pk, 0, &pk_len},
-    {2, SQL_C_DOUBLE, &data, 0, &data_len}
-  };
-
-  string insert_string{};
-  string comma{};
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    insert_string += comma + "(" + std::to_string(i) + "," + INSERTED_DATA[i] + ")";
-    comma = ",";
-  }
-
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table and assert affected rows
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, insert_string));
-
-  rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(affected_rows, NUM_OF_INSERTS);
-
-  odbcHandler.CloseStmt();
-
-  // Select all from the tables and assert that the following attributes of the type is correct:
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-
-  // Make sure inserted values are correct
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    rcode = SQLFetch(odbcHandler.GetStatementHandle()); // retrieve row-by-row
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-    ASSERT_EQ(pk, i);
-    if (INSERTED_DATA[i] != "NULL") {
-      ASSERT_EQ(data_len, DOUBLE_BYTES_EXPECTED);
-      ASSERT_EQ(data, StringToDouble(INSERTED_DATA[i]));
-    }
-    else {
-      ASSERT_EQ(data_len, SQL_NULL_DATA);
-    }
-  }
-
-  // Assert that there is no more data
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_Money, Insertion_Fail) {
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<string> INVALID_INSERTED_VALUES = {
-    "AAA",
-    "-922337203685477.5809",
-    "922337203685477.5808",
-    "999999999999999999999999999999",
-  };
-  const int NUM_OF_INSERTS = INVALID_INSERTED_VALUES.size();
-
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Attempt to insert values that are out of range and assert that they all fail
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    string insert_string = "(" + std::to_string(i) + "," + INVALID_INSERTED_VALUES[i] + ")";
-
-    rcode = SQLExecDirect(odbcHandler.GetStatementHandle(), (SQLCHAR *)InsertStatement(TABLE_NAME, insert_string).c_str(), SQL_NTS);
-    ASSERT_EQ(rcode, SQL_ERROR);
-  }
-
-  // Select all from the tables and assert that nothing was inserted
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_Money, Update_Success) {
-  const string PK_INSERTED = "0";
-  const string DATA_INSERTED = "0";
-
-  const vector <string> DATA_UPDATED_VALUES = {
-    "-922337203685477.5808",
-    "922337203685477.5807",
-    "0"
-  };
-  const int NUM_OF_DATA = DATA_UPDATED_VALUES.size();
-
-  const string INSERT_STRING = "(" + PK_INSERTED + "," + DATA_INSERTED + ")";
-  const string UPDATE_WHERE_CLAUSE = COL1_NAME + " = " + PK_INSERTED;
-
-  const int AFFECTED_ROWS_EXPECTED = 1;
-
-  int pk;
-  double data;
-  SQLLEN pk_len;
-  SQLLEN data_len;
-  SQLLEN affected_rows;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {
-    {1, SQL_C_LONG, &pk, 0, &pk_len},
-    {2, SQL_C_DOUBLE, &data, 0, &data_len}
-  };
-
-  vector<pair<string, string>> update_col{};
-
-  for (int i = 0; i < NUM_OF_DATA; i++) {
-    update_col.push_back(pair<string, string>(COL2_NAME, DATA_UPDATED_VALUES[i]));
-  }
-
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table using the correct ODBC data type mapping.
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, INSERT_STRING));
-  odbcHandler.CloseStmt();
-
-  // Bind Columns
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  // Assert that value is inserted properly
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-  ASSERT_EQ(pk, atoi(PK_INSERTED.c_str()));
-  ASSERT_EQ(data_len, DOUBLE_BYTES_EXPECTED);
-  ASSERT_EQ(data, StringToDouble(DATA_INSERTED));
-
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-  odbcHandler.CloseStmt();
-
-  // Update value multiple times
-  for (int i = 0; i < NUM_OF_DATA; i++) {
-    odbcHandler.ExecQuery(UpdateTableStatement(TABLE_NAME, vector<pair<string, string>>{update_col[i]}, UPDATE_WHERE_CLAUSE));
-
-    rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    ASSERT_EQ(affected_rows, AFFECTED_ROWS_EXPECTED);
-
-    odbcHandler.CloseStmt();
-
-    // Assert that updated value is present
-    odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-    rcode = SQLFetch(odbcHandler.GetStatementHandle());
-
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-    ASSERT_EQ(pk, atoi(PK_INSERTED.c_str()));
-    ASSERT_EQ(data_len, DOUBLE_BYTES_EXPECTED);
-    ASSERT_EQ(data, StringToDouble(DATA_UPDATED_VALUES[i]));
-
-    rcode = SQLFetch(odbcHandler.GetStatementHandle());
-    ASSERT_EQ(rcode, SQL_NO_DATA);
-    odbcHandler.CloseStmt();
-  }
-
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_Money, Update_Fail) {
-  const string PK_INSERTED = "0";
-  const string DATA_INSERTED = "0";
-  const string DATA_UPDATED_VALUE = "999999999999999999999999999999";
-
-  const string INSERT_STRING = "(" + PK_INSERTED + "," + DATA_INSERTED + ")";
-  const string UPDATE_WHERE_CLAUSE = COL1_NAME + " = " + PK_INSERTED;
-
-  const int AFFECTED_ROWS_EXPECTED = 1;
-
-  int pk;
-  double data;
-  SQLLEN pk_len;
-  SQLLEN data_len;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {
-    {1, SQL_C_LONG, &pk, 0, &pk_len},
-    {2, SQL_C_DOUBLE, &data, 0, &data_len}
-  };
-
-  const vector<pair<string, string>> UPDATE_COL = {
-    {COL2_NAME, DATA_UPDATED_VALUE}
-  };
-
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table using the correct ODBC data type mapping.
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, INSERT_STRING));
-  odbcHandler.CloseStmt();
-
-  // Bind Columns
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  // Assert that value is inserted properly
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-  ASSERT_EQ(pk, atoi(PK_INSERTED.c_str()));
-  ASSERT_EQ(data_len, DOUBLE_BYTES_EXPECTED);
-  ASSERT_EQ(data, StringToDouble(DATA_INSERTED));
-
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-  odbcHandler.CloseStmt();
-
-  // Update value and assert an error is present
-  rcode = SQLExecDirect(odbcHandler.GetStatementHandle(), (SQLCHAR *)UpdateTableStatement(TABLE_NAME, UPDATE_COL, UPDATE_WHERE_CLAUSE).c_str(), SQL_NTS);
-  ASSERT_EQ(rcode, SQL_ERROR);
-  odbcHandler.CloseStmt();
-
-  // Assert that no values changed
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-  ASSERT_EQ(pk, atoi(PK_INSERTED.c_str()));
-  ASSERT_EQ(data_len, DOUBLE_BYTES_EXPECTED);
-  ASSERT_EQ(data, StringToDouble(DATA_INSERTED));
-
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_Money, Arithmetic_Operators) {
-  const vector<pair<string, string>> TABLE_COLUMNS = {
-    {COL1_NAME, DATATYPE_NAME + " PRIMARY KEY"},
-    {COL2_NAME, DATATYPE_NAME}
-  };
-
-  double pk;
-  double data;
-  SQLLEN pk_len;
-  SQLLEN data_len;
-  SQLLEN affected_rows;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<string> INSERTED_PK = {
-    "-987654321.1234",
-    "-123456789.4321",
-    "0"
-  };
-
-  const vector<string> INSERTED_DATA = {
-    "2.01",
-    "-100000000.01",
-    "922337203685477.5807"
-  };
-  const int NUM_OF_DATA = INSERTED_DATA.size();
-
-  const vector<string> OPERATIONS_QUERY = {
-    COL1_NAME + "+" + COL2_NAME,
-    COL1_NAME + "-" + COL2_NAME,
-    COL1_NAME + "/" + COL2_NAME,
-    COL1_NAME + "*" + COL2_NAME
-  };
-  const int NUM_OF_OPERATIONS = OPERATIONS_QUERY.size();
-
-  vector<vector<double>> expected_results = {};
-
-  // initialization of expected_results
-  for (int i = 0; i < INSERTED_DATA.size(); i++) {
-    expected_results.push_back({});
-    const double data_1 = StringToDouble(INSERTED_PK[i]);
-    const double data_2 = StringToDouble(INSERTED_DATA[i]);
-
-    expected_results[i].push_back(data_1 + data_2);
-    expected_results[i].push_back(data_1 - data_2);
-    expected_results[i].push_back(data_1 / data_2);
-    expected_results[i].push_back(data_1 * data_2);
-  }
-
-  double col_results[NUM_OF_OPERATIONS];
-  SQLLEN col_len[NUM_OF_OPERATIONS];
-  vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {};
-
-  // initialization for BIND_COLUMNS
-  for (int i = 0; i < NUM_OF_OPERATIONS; i++) {
-    tuple<int, int, SQLPOINTER, int, SQLLEN *> tuple_to_insert(i + 1, SQL_C_DOUBLE, (SQLPOINTER)&col_results[i], 0, &col_len[i]);
-    BIND_COLUMNS.push_back(tuple_to_insert);
-  }
-
-  string insert_string{};
-  string comma{};
-
-  // insert_string initialization
-  for (int i = 0; i < INSERTED_DATA.size(); i++) {
-    insert_string += comma + "(" + INSERTED_PK[i] + "," + INSERTED_DATA[i] + ")";
-    comma = ",";
-  }
-
-  // Create table
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table and assert affected rows
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, insert_string));
-
-  rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(affected_rows, INSERTED_DATA.size());
-
-  // Make sure inserted values are correct and operations
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, OPERATIONS_QUERY, vector<string>{}) + " ORDER BY " + COL1_NAME);
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  for (int i = 0; i < INSERTED_DATA.size(); i++) {
-    rcode = SQLFetch(odbcHandler.GetStatementHandle());
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-
-    for (int j = 0; j < NUM_OF_OPERATIONS; j++) {
-      ASSERT_EQ(col_len[j], DOUBLE_BYTES_EXPECTED);
-      ASSERT_EQ(col_results[j], expected_results[i][j]);
-    }
-  }
-
-  // Assert that there is no more data
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_Money, Arithmetic_Functions) {
-  const vector<pair<string, string>> TABLE_COLUMNS = {
-    {COL1_NAME, DATATYPE_NAME + " PRIMARY KEY"},
-    {COL2_NAME, DATATYPE_NAME}
-  };
-
-  double pk;
-  double data;
-  SQLLEN pk_len;
-  SQLLEN data_len;
-  SQLLEN affected_rows;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<string> INSERTED_PK = {
-    "-987654321.1234",
-    "-123456789.4321",
-    "0"
-  };
-
-  vector<string> INSERTED_DATA = {
-    "2.01",
-    "-100000000.01",
-    "922337203685477.5807"
-  };
-  const int NUM_OF_DATA = INSERTED_DATA.size();
-
-  const vector<string> OPERATIONS_QUERY = {
-    "MIN(" + COL1_NAME + ")",
-    "MAX(" + COL1_NAME + ")",
-    "SUM(" + COL1_NAME + ")",
-    "AVG(" + COL1_NAME + ")"
-  };
-  const int NUM_OF_OPERATIONS = OPERATIONS_QUERY.size();
-
-  // initialization of expected_results
-  vector<double> expected_results = {};
-  double curr = StringToDouble(INSERTED_PK[0]);
-  double min_expected = curr, max_expected = curr, sum_expected = curr, avg_expected = 0;
-  for (int i = 1; i < INSERTED_PK.size(); i++) {
-    curr = StringToDouble(INSERTED_PK[i]);
-    sum_expected += curr;
-    min_expected = std::min(min_expected, curr);
-    max_expected = std::max(max_expected, curr);
-  }
-  avg_expected = sum_expected / INSERTED_PK.size();
-  expected_results.push_back(min_expected);
-  expected_results.push_back(max_expected);
-  expected_results.push_back(sum_expected);
-  expected_results.push_back(avg_expected);
-
-  double col_results[NUM_OF_OPERATIONS];
-  SQLLEN col_len[NUM_OF_OPERATIONS];
-  vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {};
-
-  // initialization for bind_columns
-  for (int i = 0; i < NUM_OF_OPERATIONS; i++) {
-    tuple<int, int, SQLPOINTER, int, SQLLEN *> tuple_to_insert(i + 1, SQL_C_DOUBLE, (SQLPOINTER)&col_results[i], 0, &col_len[i]);
-    BIND_COLUMNS.push_back(tuple_to_insert);
-  }
-
-  string insert_string{};
-  string comma{};
-
-  // insert_string initialization
-  for (int i = 0; i < INSERTED_DATA.size(); i++) {
-    insert_string += comma + "(" + INSERTED_PK[i] + "," + INSERTED_DATA[i] + ")";
-    comma = ",";
-  }
-
-  // Create table
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table and assert affected rows
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, insert_string));
-
-  rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(affected_rows, INSERTED_DATA.size());
-  odbcHandler.CloseStmt();
-
-  // Make sure inserted values are correct and operations  
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, OPERATIONS_QUERY, vector<string>{}));
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  for (int i = 0; i < NUM_OF_OPERATIONS; i++) {
-    ASSERT_EQ(col_len[i], DOUBLE_BYTES_EXPECTED);
-    ASSERT_EQ(col_results[i], expected_results[i]);
-  }
-
-  // Assert that there is no more data
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_Money, View_Creation) {
-  const string VIEW_QUERY = "SELECT * FROM " + TABLE_NAME;
-
-  int pk;
-  double data;
-  SQLLEN pk_len;
-  SQLLEN data_len;
-  SQLLEN affected_rows;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<string> INSERTED_DATA = {
-    "123456789.0001",
-    "1",
-    "NULL"
-  };
-  const int NUM_OF_INSERTS = INSERTED_DATA.size();
-
-  const vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {
-    {1, SQL_C_LONG, &pk, 0, &pk_len},
-    {2, SQL_C_DOUBLE, &data, 0, &data_len}
-  };
-
-  string insert_string{};
-  string comma{};
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    insert_string += comma + "(" + std::to_string(i) + "," + INSERTED_DATA[i] + ")";
-    comma = ",";
-  }
-
-  // Create Table
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table and assert affected rows
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, insert_string));
-
-  rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(affected_rows, NUM_OF_INSERTS);
-
-  odbcHandler.CloseStmt();
-
-  // Create view
-  odbcHandler.ExecQuery(CreateViewStatement(VIEW_NAME, VIEW_QUERY));
-  odbcHandler.CloseStmt();
-
-  // Select all from the tables and assert that the following attributes of the type is correct:
-  odbcHandler.ExecQuery(SelectStatement(VIEW_NAME, {"*"}, vector<string>{COL1_NAME}));
-
-  // Make sure inserted values are correct
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    rcode = SQLFetch(odbcHandler.GetStatementHandle()); // retrieve row-by-row
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-    ASSERT_EQ(pk, i);
-
-    if (INSERTED_DATA[i] != "NULL") {
-      ASSERT_EQ(data_len, DOUBLE_BYTES_EXPECTED);
-      ASSERT_EQ(data, StringToDouble(INSERTED_DATA[i]));
-    }
-    else {
-      ASSERT_EQ(data_len, SQL_NULL_DATA);
-    }      
-  }
-
-  // Assert that there is no more data
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("VIEW", VIEW_NAME));
-
-  odbcHandler.CloseStmt();
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_Money, Table_Unique_Constraints) {
-  const vector<pair<string, string>> TABLE_COLUMNS = {
-    {COL1_NAME, "INT PRIMARY KEY"},
-    {COL2_NAME, DATATYPE_NAME + " UNIQUE"}
-  };
-  const string UNIQUE_COLUMN_NAME = COL2_NAME;
-
-  int pk;
-  double data;
-  SQLLEN pk_len;
-  SQLLEN data_len;
-  SQLLEN affected_rows;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<string> INSERTED_DATA = {
-    "1000.0001",
-    "-1000.0001"
-  };
-  const int NUM_OF_INSERTS = INSERTED_DATA.size();
-
-  string insert_string{};
-  string comma{};
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    insert_string += comma + "(" + std::to_string(i) + "," + INSERTED_DATA[i] + ")";
-    comma = ",";
-  }
-
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS));
-  odbcHandler.CloseStmt();
-
-  // Check if unique constraint still matches after creation
-  char column_name[BUFFER_SIZE];
-  char type_name[BUFFER_SIZE];
-
-  const vector<tuple<int, int, SQLPOINTER, int>> TABLE_BIND_COLUMNS = {
-    {1, SQL_C_CHAR, column_name, BUFFER_SIZE},
-  };
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(TABLE_BIND_COLUMNS));
-
-  const string UNIQUE_KEY_QUERY =
-    "SELECT C.COLUMN_NAME FROM "
-    "INFORMATION_SCHEMA.TABLE_CONSTRAINTS T "
-    "JOIN INFORMATION_SCHEMA.CONSTRAINT_COLUMN_USAGE C "
-    "ON C.CONSTRAINT_NAME=T.CONSTRAINT_NAME "
-    "WHERE "
-    "C.TABLE_NAME='" +
-    TABLE_NAME.substr(TABLE_NAME.find('.') + 1, TABLE_NAME.length()) + "' "
-    "AND T.CONSTRAINT_TYPE='UNIQUE'";
-  odbcHandler.ExecQuery(UNIQUE_KEY_QUERY);
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(string(column_name), UNIQUE_COLUMN_NAME);
-
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  EXPECT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table and assert affected rows
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, insert_string));
-
-  rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(affected_rows, NUM_OF_INSERTS);
-
-  odbcHandler.CloseStmt();
-
-  // Select all from the tables and assert that the following attributes of the type is correct:
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-
-  // Make sure inserted values are correct
-  const vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {
-    {1, SQL_C_LONG, &pk, 0, &pk_len},
-    {2, SQL_C_DOUBLE, &data, 0, &data_len}
-  };
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    rcode = SQLFetch(odbcHandler.GetStatementHandle()); // retrieve row-by-row
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-    ASSERT_EQ(pk, i);
-    if (INSERTED_DATA[i] != "NULL") {
-      ASSERT_EQ(data_len, DOUBLE_BYTES_EXPECTED);
-      ASSERT_EQ(data, StringToDouble(INSERTED_DATA[i]));
-    }
-    else {
-      ASSERT_EQ(data_len, SQL_NULL_DATA);
-    }
-  }
-
-  // Assert that there is no more data
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-
-  // Attempt to insert
-  const vector<string> INVALID_INSERTED_VALUES = {
-    "1000.0001",
-    "-1000.0001"
-  };
-  const int NUM_OF_INVALID = INVALID_INSERTED_VALUES.size();
-
-  // Attempt to insert values that violates unique constraint and assert that they all fail
-  for (int i = NUM_OF_INVALID; i < 2 * NUM_OF_INVALID; i++) {
-    string insert_string = "(" + std::to_string(i) + "," + INVALID_INSERTED_VALUES[i - NUM_OF_INVALID] + ")";
-
-    rcode = SQLExecDirect(odbcHandler.GetStatementHandle(), (SQLCHAR *)InsertStatement(TABLE_NAME, insert_string).c_str(), SQL_NTS);
-    ASSERT_EQ(rcode, SQL_ERROR);
-  }
-
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
-
-TEST_F(PSQL_DataTypes_Money, Table_Composite_Keys) {
-  const vector<pair<string, string>> TABLE_COLUMNS = {
-    {COL1_NAME, "INT"},
-    {COL2_NAME, DATATYPE_NAME}
-  };
-  const string PKTABLE_NAME = TABLE_NAME.substr(TABLE_NAME.find('.') + 1, TABLE_NAME.length());
-  const string SCHEMA_NAME = TABLE_NAME.substr(0, TABLE_NAME.find('.'));
-
-  const vector<string> PK_COLUMNS = {
-    COL1_NAME,
-    COL2_NAME
-  };
-
-  string table_constraints{"PRIMARY KEY ("};
-  string comma{};
-  for (int i = 0; i < PK_COLUMNS.size(); i++) {
-    table_constraints += comma + PK_COLUMNS[i];
-    comma = ",";
-  }
-  table_constraints += ")";
-
-  int pk;
-  double data;
-  SQLLEN pk_len;
-  SQLLEN data_len;
-  SQLLEN affected_rows;
-
-  RETCODE rcode;
-  OdbcHandler odbcHandler(Drivers::GetDriver(ServerType::PSQL));
-
-  const vector<string> INSERTED_DATA = {
-    "1000.01",
-    "-100.0001"
-  };
-  const int NUM_OF_INSERTS = INSERTED_DATA.size();
-
-  const vector<tuple<int, int, SQLPOINTER, int, SQLLEN *>> BIND_COLUMNS = {
-    {1, SQL_C_LONG, &pk, 0, &pk_len},
-    {2, SQL_C_DOUBLE, &data, 0, &data_len}
-  };
-
-  string insert_string{};
-  comma = "";
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    insert_string += comma + "(" + std::to_string(i) + "," + INSERTED_DATA[i] + ")";
-    comma = ",";
-  }
-
-  odbcHandler.ConnectAndExecQuery(CreateTableStatement(TABLE_NAME, TABLE_COLUMNS, table_constraints));
-  odbcHandler.CloseStmt();
-
-  // Check if composite key still matches after creation
-  char table_name[BUFFER_SIZE];
-  char column_name[BUFFER_SIZE];
-  int key_sq{};
-  char pk_name[BUFFER_SIZE];
-
-  vector<tuple<int, int, SQLPOINTER, int>> constraints_bind_columns = {
-    {3, SQL_C_CHAR, table_name, BUFFER_SIZE},
-    {4, SQL_C_CHAR, column_name, BUFFER_SIZE},
-    {5, SQL_C_ULONG, &key_sq, BUFFER_SIZE},
-    {6, SQL_C_CHAR, pk_name, BUFFER_SIZE}
-  };
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(constraints_bind_columns));
-
-  rcode = SQLPrimaryKeys(odbcHandler.GetStatementHandle(), NULL, 0, (SQLCHAR *)SCHEMA_NAME.c_str(), SQL_NTS, (SQLCHAR *)PKTABLE_NAME.c_str(), SQL_NTS);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-
-  int curr_sq{0};
-  for (auto columnName : PK_COLUMNS) {
-    ++curr_sq;
-    rcode = SQLFetch(odbcHandler.GetStatementHandle());
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-
-    ASSERT_EQ(string(table_name), PKTABLE_NAME);
-    ASSERT_EQ(string(column_name), columnName);
-    ASSERT_EQ(key_sq, curr_sq);
-  }
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-  odbcHandler.CloseStmt();
-
-  // Insert valid values into the table and assert affected rows
-  odbcHandler.ExecQuery(InsertStatement(TABLE_NAME, insert_string));
-
-  rcode = SQLRowCount(odbcHandler.GetStatementHandle(), &affected_rows);
-  ASSERT_EQ(rcode, SQL_SUCCESS);
-  ASSERT_EQ(affected_rows, NUM_OF_INSERTS);
-
-  odbcHandler.CloseStmt();
-
-  // Select all from the tables and assert that the following attributes of the type is correct:
-  odbcHandler.ExecQuery(SelectStatement(TABLE_NAME, {"*"}, vector<string>{COL1_NAME}));
-
-  // Make sure inserted values are correct
-  ASSERT_NO_FATAL_FAILURE(odbcHandler.BindColumns(BIND_COLUMNS));
-
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    rcode = SQLFetch(odbcHandler.GetStatementHandle()); // retrieve row-by-row
-    ASSERT_EQ(rcode, SQL_SUCCESS);
-    ASSERT_EQ(pk_len, INT_BYTES_EXPECTED);
-    ASSERT_EQ(pk, i);
-    if (INSERTED_DATA[i] != "NULL") {
-      ASSERT_EQ(data_len, DOUBLE_BYTES_EXPECTED);
-      ASSERT_EQ(data, StringToDouble(INSERTED_DATA[i]));
-    }
-    else {
-      ASSERT_EQ(data_len, SQL_NULL_DATA);
-    }
-  }
-
-  // Assert that there is no more data
-  rcode = SQLFetch(odbcHandler.GetStatementHandle());
-  ASSERT_EQ(rcode, SQL_NO_DATA);
-
-  odbcHandler.CloseStmt();
-
-  // Attempt to insert values that violates composite constraint and assert that they all fail
-  for (int i = 0; i < NUM_OF_INSERTS; i++) {
-    insert_string += comma + "(" + std::to_string(i) + "," + INSERTED_DATA[i] + ")";
-    comma = ",";
-  }
-
-  rcode = SQLExecDirect(odbcHandler.GetStatementHandle(), (SQLCHAR *)InsertStatement(TABLE_NAME, insert_string).c_str(), SQL_NTS);
-  ASSERT_EQ(rcode, SQL_ERROR);
-
-  odbcHandler.ExecQuery(DropObjectStatement("TABLE", TABLE_NAME));
-}
